+++ packages/@expo/cli/src/start/server/metro/resolveFromProject.ts
@@ -41,13 +41,6 @@ export function importMetroHmrServerFromProject(
   return importFromProject(projectRoot, 'metro/src/HmrServer');
 }
 
-/** Import `@expo/metro-config` from the project. */
-export function importExpoMetroConfigFromProject(
-  projectRoot: string
-): typeof import('@expo/metro-config') {
-  return importFromProject(projectRoot, '@expo/metro-config');
-}
-
 /** Import `metro-resolver` from the project. */
 export function importMetroResolverFromProject(
   projectRoot: string
